Query generator should not treat quoted phrases specially.
Quoted strings in CQL only have syntactic meaning for the parser, they don't carry semantic meaning for the resulting query.

As a consequence of this change, they have to be used in their unquoted form (`tok2.getString()` instead of `arg2.toString()`).

NOTE: The `cql.unmasked` relation modifier may be implemented in order to support the previous behaviour of "protected" phrases.